Encounter Medicines tab UI
