+++ src/components/Medicine/MedicationsTable.tsx
@@ -1,0 +1,135 @@
+import { useTranslation } from "react-i18next";
+
+import {
+  Table,
+  TableBody,
+  TableCell,
+  TableHead,
+  TableHeader,
+  TableRow,
+} from "@/components/ui/table";
+
+import { reverseFrequencyOption } from "@/components/Questionnaire/QuestionTypes/MedicationRequestQuestion";
+
+import {
+  MEDICATION_REQUEST_TIMING_OPTIONS,
+  MedicationRequestDosageInstruction,
+  MedicationRequestRead,
+} from "@/types/emr/medicationRequest";
+
+function getFrequencyDisplay(
+  timing?: MedicationRequestDosageInstruction["timing"],
+) {
+  if (!timing) return undefined;
+  const code = reverseFrequencyOption(timing);
+  if (!code) return undefined;
+  return {
+    code,
+    meaning: MEDICATION_REQUEST_TIMING_OPTIONS[code].display,
+  };
+}
+
+// Helper function to format dosage in Rx style
+function formatDosage(instruction: MedicationRequestDosageInstruction) {
+  if (!instruction.dose_and_rate) return "";
+
+  if (instruction.dose_and_rate.type === "calculated") {
+    const { dose_range } = instruction.dose_and_rate;
+    if (!dose_range) return "";
+    return `${dose_range.low.value}${dose_range.low.unit.display} - ${dose_range.high.value}${dose_range.high.unit.display}`;
+  }
+
+  const { dose_quantity } = instruction.dose_and_rate;
+  if (!dose_quantity?.value || !dose_quantity.unit) return "";
+
+  return `${dose_quantity.value} ${dose_quantity.unit.display}`;
+}
+
+// Helper function to format dosage instructions in Rx style
+function formatSig(instruction: MedicationRequestDosageInstruction) {
+  const parts: string[] = [];
+
+  // Add route if present
+  if (instruction.route?.display) {
+    parts.push(`Via ${instruction.route.display}`);
+  }
+
+  // Add method if present
+  if (instruction.method?.display) {
+    parts.push(`by ${instruction.method.display}`);
+  }
+
+  // Add site if present
+  if (instruction.site?.display) {
+    parts.push(`to ${instruction.site.display}`);
+  }
+
+  return parts.join(" ");
+}
+
+interface MedicationsTableProps {
+  medications: MedicationRequestRead[];
+}
+
+export const MedicationsTable = ({ medications }: MedicationsTableProps) => {
+  const { t } = useTranslation();
+
+  if (!medications?.length) {
+    return (
+      <div className="flex h-[200px] items-center justify-center rounded-lg border-2 border-dashed p-4 text-muted-foreground">
+        {t("no_medications_found_for_this_encounter")}
+      </div>
+    );
+  }
+
+  return (
+    <div className="border rounded-lg overflow-hidden">
+      <Table>
+        <TableHeader>
+          <TableRow className="divide-x bg-gray-100">
+            <TableHead>{t("medicine")}</TableHead>
+            <TableHead>{t("dosage")}</TableHead>
+            <TableHead>{t("frequency")}</TableHead>
+            <TableHead>{t("duration")}</TableHead>
+            <TableHead>{t("instructions")}</TableHead>
+          </TableRow>
+        </TableHeader>
+        <TableBody>
+          {medications.map((medication) => {
+            const instruction = medication.dosage_instruction[0];
+            const frequency = getFrequencyDisplay(instruction?.timing);
+            const dosage = formatDosage(instruction);
+            const duration = instruction?.timing?.repeat?.bounds_duration;
+            const remarks = formatSig(instruction);
+            const notes = medication.note;
+            return (
+              <TableRow key={medication.id} className="divide-x font-medium">
+                <TableCell className="py-2 px-3">
+                  {medication.medication?.display}
+                </TableCell>
+                <TableCell className="py-2 px-3">{dosage}</TableCell>
+                <TableCell className="py-2 px-3">
+                  {instruction?.as_needed_boolean
+                    ? `${t("as_needed_prn")} (${instruction?.as_needed_for?.display})`
+                    : frequency?.meaning}
+                  {instruction?.additional_instruction?.[0]?.display && (
+                    <div className="text-sm text-gray-600">
+                      {instruction.additional_instruction[0].display}
+                    </div>
+                  )}
+                </TableCell>
+                <TableCell className="py-2 px-3">
+                  {duration ? `${duration.value} ${duration.unit}` : "-"}
+                </TableCell>
+                <TableCell className="py-2 px-3">
+                  {remarks || "-"}
+                  {notes ? ` (${t("note")}: ${notes})` : ""}
+                </TableCell>
+              </TableRow>
+            );
+          })}
+        </TableBody>
+      </Table>
+    </div>
+  );
+};
